exp/services/webauth: rename commands package to cmd
