+++ exp/services/webauth/cmd/genjwtkey.go
@@ -1,4 +1,4 @@
-package commands
+package cmd
 
 import (
 	"github.com/spf13/cobra"
